linalg & fem: introduce and adjust testing
